fix: conditionalize dracut services for anaconda
This patch contains a fix for the enablement of
coreos-enable-network.service - it used to depend on the initrd.target
which was just completely wrong and just a bug/error. Fixing that makes
the system work with anaconda again.
Added a generator to make sure the setup user service doesn't depend on
the boot device if we're not gonna run ignition.
Also, just rewrote the Makefile to make sure it's easier to debug in the
future.

Signed-off-by: Antonio Murdaca <antoniomurdaca@gmail.com>